+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/preference/TikTokPreferenceFragment.java
@@ -296,6 +296,9 @@ public class TikTokPreferenceFragment extends AbstractPreferenceFragment {
                     SettingsStatus.authorRegionEnabled && Settings.SHOW_AUTHOR_REGION.get(),
                     SettingsStatus.videoOverlaysEnabled && Settings.HIDE_FEED_CAPTION.get(),
                     SettingsStatus.videoOverlaysEnabled && Settings.HIDE_FEED_MUSIC.get(),
+                    SettingsStatus.videoOverlaysEnabled && Settings.HIDE_FEED_ACTION_BAR.get(),
+                    SettingsStatus.videoOverlaysEnabled && Settings.HIDE_FEED_SURVEYS.get(),
+                    SettingsStatus.videoOverlaysEnabled && Settings.HIDE_STATUS_BAR.get(),
                     SettingsStatus.videoOverlaysEnabled && Settings.HIDE_VISUAL_SEARCH.get(),
                     (SettingsStatus.videoOverlaysEnabled || SettingsStatus.hideFeedLiveButtonEnabled)
                             && Settings.HIDE_LIVE_ENTRANCE.get(),